+++ examples/advanced/src/gen/clients/axios/petService/addFiles.ts
@@ -25,6 +25,8 @@ export async function addFiles(
       const value = requestData[key as keyof typeof requestData]
       if (typeof value === 'string' || (value as unknown) instanceof Blob) {
         formData.append(key, value as unknown as string | Blob)
+      } else {
+        formData.append(key, JSON.stringify(value))
       }
     })
   }