+++ examples/client/src/gen/clients/axios/petService/uploadFile.ts
@@ -37,6 +37,8 @@ export async function uploadFile(
       const value = requestData[key as keyof typeof requestData]
       if (typeof value === 'string' || (value as unknown) instanceof Blob) {
         formData.append(key, value as unknown as string | Blob)
+      } else {
+        formData.append(key, JSON.stringify(value))
       }
     })
   }