+++ packages/plugin-client/src/components/Client.tsx
@@ -194,6 +194,8 @@ export function Client({
       const value = requestData[key as keyof typeof requestData];
       if (typeof value === 'string' || (value as unknown) instanceof Blob) {
         formData.append(key, value as unknown as string | Blob);
+      } else {
+        formData.append(key, JSON.stringify(value));
       }
     })
    }